Split up functions from VerifiableCredentialManager
